requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
aiohttp==3.9.1

# Machine Learning
torch==2.1.1
transformers==4.36.2
scikit-learn==1.3.2
huggingface-hub==0.19.4
joblib==1.3.2
xgboost==2.0.2

# Analytics
plotly==5.17.0
dash==2.14.2
streamlit==1.28.2

# Web scraping
beautifulsoup4==4.12.2
selenium==4.15.2

# Monitoring
prometheus-client==0.19.0
loguru==0.7.2

# Async support
asyncpg==0.29.0
asyncio-mqtt==0.16.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
